+++ packages/biscuit-store/tests/container.test.js
@@ -18,3 +18,23 @@ it('container test', () => {
 	expect(typeof start.dispatch).toEqual('function');
 	expect(typeof step.dispatch).toEqual('function');
 });
+
+it('container test', () => {
+	const { actions } = createStore({
+		name: 'test2',
+		initial: { value: 1 },
+		actions: {
+			start: 'start/action',
+			step: 'step/action',
+		},
+		strictMode: false,
+		addToСontainer: true,
+	});
+
+	const { start, step } = container.extract('test2');
+
+	expect(typeof actions.start.dispatch).toEqual('function');
+	expect(typeof actions.step.dispatch).toEqual('function');
+	expect(typeof start.dispatch).toEqual('function');
+	expect(typeof step.dispatch).toEqual('function');
+});