+++ packages/types/types/store.ts
@@ -78,6 +78,7 @@ export interface StoreParams <A extends StateAction, I> {
 	actions: { [propName: string]: A };
 }
 
+
 /**
  * The interface describes
  * a set of createStore options
@@ -98,6 +99,11 @@ export interface StoreSettings {
 	 * to the store during initialization
 	 */
 	initialCall?: <T>(store?: T) => void;
+	/** Combined actions */
+	combineActions?: {
+		[propName: string]:
+		<S = object, P = object>(state?: S, payload?: P) => void;
+	};
 	/** Enable or disable strict mode */
 	strictMode?: boolean;
 }
@@ -136,3 +142,13 @@ export interface Container {
 	 */
 	extract: (storeName: string) => {[propName: string]: StateAction};
 }
+
+/** The type describes the arguments of the combineActions method */
+export type CombineProto<S = object> =
+	{ [propName: string]: <P = object>(state?: S, payload?: P) => void };
+
+/** The interface describes the returned combineActions parameters */
+export interface CombineActions {
+	middleware: Middleware[];
+	actions: Actions;
+};