+++ CHANGELOG.md
@@ -1,4 +1,19 @@
-# @biscuit-store
+# @biscuit-store/core
+
+## 1.2.0
+### Major Changes
+
+- added combineActions function
+- added addToContainer field to createStore
+- added checking the existence of a repository by name
+- added combineAction field to createStore
+- refactor creator
+- added test for new functions
+
+## 1.1.0
+### Major Changes
+- added container function
+- added initialCall  function
 
 ## 1.0.0
 ### Major Changes
@@ -13,17 +28,74 @@
 - ec22884: remove combineStateCollection
 - 79d022f: fix docs
 
-### Minor Changes
+## 0.9.99
+### Patch Changes
 
-- ec22884: added stateCollection.test.js
-- 038a474: fix debounce function
-- 442379a: refactor compareObject helper function
-- 7cca360: Remove manager.clone()
+- Added adapter unit test
+- Added call method to adapter
+- Added channel method to adapter
+- Added createActionTo unit test
+- refactor store adn actions
+
+# @biscuit-store/adapter
+
+## 1.1.0
+### Major Changes
+
+- State and payload include to context
+- Added includeContext function
+
+## 0.9.99
+### Patch Changes
+
+- Added adapter unit test
+- Added call method to adapter
+- Added channel method to adapter
+- Added createActionTo unit test
+- refactor store adn actions
+
+# @biscuit-store/react
+
+## 2.0.0
+### Patch Changes
+
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+  - @biscuit-store/core@2.0.0
+
+## 1.1.0
+### Major Changes
+
+- Added listen function
+- Refactor objerver function
 
 ### Patch Changes
 
-- 7cca360: Fix manager.merge()
-- 7cca360: Department of the manager module
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+  - @biscuit-store/core@2.0.0
+
+## 1.0.0
+### Major Changes
+
+- 038a474: fix useDebounce hook
+- 038a474: fix unsubscribe to susscribe decorator
+- 038a474: add subscribe test
+- 038a474: add useDispatchDebounce hook
+- 038a474: add observer test
+- 038a474: add useSubscribe test
+- 038a474: add useDispatchThrottle test
+- 038a474: add useDispatch unit test
+
+### Patch Changes
+  - @biscuit-store/core@1.0.0
+
+## 1.0.0
+### Patch Changes
+  - @biscuit-store/core@1.0.0
 
 ## 0.9.99
 ### Patch Changes